cln-rpc: The JSON `number` type is a float not an int
Got some issues parsing dates for example.

+++ contrib/msggen/msggen/rust.py
@@ -35,7 +35,7 @@ typemap = {
     'msat': 'Amount',
     'msat_or_all': 'AmountOrAll',
     'msat_or_any': 'AmountOrAny',
-    'number': 'i64',
+    'number': 'f64',
     'pubkey': 'Pubkey',
     'short_channel_id': 'ShortChannelId',
     'signature': 'String',
